requirements.txt
certifi==2018.11.29
chardet==3.0.4
durations==0.3.3
idna==2.8
python-dateutil==2.7.5
requests==2.21.0
schedule==0.5.0
six==1.12.0
urllib3==1.24.1
